+++ libvast/src/system/connect_to_node.cpp
@@ -43,8 +43,9 @@ caf::expected<actor>
 connect_to_node(scoped_actor& self, const caf::settings& opts) {
   // Fetch values from config.
   auto id = get_or(opts, "system.node-id", defaults::system::node_id);
-  auto dir = get_or(opts, "system.directory", defaults::system::directory);
-  auto abs_dir = path{dir}.complete();
+  auto db_dir
+    = get_or(opts, "system.db-directory", defaults::system::db_directory);
+  auto abs_dir = path{db_dir}.complete();
   endpoint node_endpoint;
   auto str = get_or(opts, "system.endpoint", defaults::system::endpoint);
   if (!parsers::endpoint(str, node_endpoint))